+++ app/javascript/packs/application.js
@@ -8,9 +8,9 @@ require("@rails/ujs").start()
 require("@rails/activestorage").start()
 require("channels")
 require("../card")
-require("../search")
+require("../top_search")
 require("../shop")
-require("../search_shop")
+require("../search")
 
 // Uncomment to copy all static images under ../images to the output folder and reference
 // them with the image_pack_tag helper in views (e.g <%= image_pack_tag 'rails.png' %>)
