+++ app/views/items/index.html.erb
@@ -21,8 +21,8 @@
     <% if @items.present? %>
       <% @items.each do |item| %>
         <li class='list'>
-          <%= link_to shop_item_path(@shop.id, item.id), method: :get, class: 'shop-item' do %>
-            <%= image_tag item.image, class: 'shop-item-img' %>
+          <%= link_to shop_item_path(@shop.id, item.id), method: :get, class: 'list-item' do %>
+            <%= image_tag item.image, class: 'list-item-img' %>
             <div class='shop-item-info'>
               <h4 class='shop-item-name'><%= item.name %></h4>
               <div class='shop-item-price'>
