+++ app/views/items/search.html.erb
@@ -1,17 +1,25 @@
-<h1>商品一覧</h1>
-<%if @results.length !=0 %>
-  <% @results.each do |result| %>
-  <td>
-  <br>
-    <%= link_to shop_item_path(result.shop.id, result.id), method: :get do %>
-    <li>
-      <%= image_tag result.image.variant(resize: '400x400') %>
-      <%= result.name %>
-      <%= result.shop.name %>
-      <%= result.price %>
-    </li>
+<%= render "shared/title_header" %>
+
+<%= render "shared/header" %>
+
+<%= render "shared/search" %>
+
+<div class='search-main'>
+  <h1 class='search-list-title'><%= @results.count %>件</h1>
+  <ul class='item-lists'>
+    <%if @results.length !=0 %>
+      <% @results.each do |result| %>
+        <li class='list'>
+          <%= link_to shop_item_path(result.shop.id, result.id), method: :get, class: 'list-item' do %>
+          <%= image_tag result.image, class: 'list-item-img' %>
+          <p class='list-item-name'><%= result.name %></p>
+          <p class='list-item-price'>¥<%= result.price %></p>
+          <p class='list-shop-name'><%= result.shop.name %></p>
+        </li>
+        <% end %>
+      <% end %>
+    <% else %>
+      該当する商品はありません
     <% end %>
-  <% end %>
-<% else %>
-  該当する商品はありません
-<% end %>
+  </ul>
+</div>
